Add External Links Support
The left-hand menu previously only supported showing Resources, but often there are external resources outside of Motor Admin that also need to show up.

This commit adds a concept of an ExternalLink. It behaves similar to a resource in that it can have a name, icon, change visibility, and be sorted alongside regular resources.